Made posts popups a shared modal, and working 'view all comments' added
